Spawn gpu controllers w/ dynamic ports
