+++ pkg/api/runc.go
@@ -3,12 +3,9 @@ package api
 // Implements the task service functions for runc
 
 import (
-	"bytes"
 	"context"
 	"errors"
 	"fmt"
-	"io"
-	"os/exec"
 	"syscall"
 	"time"
 
@@ -63,15 +60,11 @@ func (s *service) RuncManage(ctx context.Context, args *task.RuncManageArgs) (*t
 	state.JobState = task.JobState_JOB_RUNNING
 	state.GPU = args.GPU
 
-	var gpuCmd *exec.Cmd
-	gpuOutBuf := &bytes.Buffer{}
 	if args.GPU {
 		log.Info().Msg("GPU support requested, assuming process was already started with LD_PRELOAD")
 		if args.GPU {
-			gpuOut := io.Writer(gpuOutBuf)
-			gpuCmd, err = s.StartGPUController(ctx, args.UID, args.GID, args.Groups, gpuOut, args.ContainerID)
+			err = s.StartGPUController(ctx, args.UID, args.GID, args.Groups, args.ContainerID)
 			if err != nil {
-				log.Error().Err(err).Str("stdout/stderr", gpuOutBuf.String()).Msg("failed to start GPU controller")
 				return nil, fmt.Errorf("failed to start GPU controller: %v", err)
 			}
 		}
@@ -106,27 +99,17 @@ func (s *service) RuncManage(ctx context.Context, args *task.RuncManageArgs) (*t
 		if err != nil {
 			log.Error().Err(err).Msg("failed to update state after done")
 		}
-		if gpuCmd != nil {
-			err = gpuCmd.Process.Kill()
-			if err != nil {
-				log.Error().Err(err).Msg("failed to kill GPU controller after runc container exit")
-			}
+		if s.GetGPUController(state.JID) != nil {
+			s.StopGPUController(state.JID)
 		}
 	}()
 
 	// Clean up GPU controller and also handle premature exit
-	if gpuCmd != nil {
+	if s.GetGPUController(state.JID) != nil {
 		s.wg.Add(1)
 		go func() {
 			defer s.wg.Done()
-			err := gpuCmd.Wait()
-			if err != nil {
-				log.Debug().Err(err).Msg("GPU controller Wait()")
-			}
-			log.Info().Int("PID", gpuCmd.Process.Pid).
-				Int("status", gpuCmd.ProcessState.ExitCode()).
-				Str("out/err", gpuOutBuf.String()).
-				Msg("GPU controller exited")
+			s.WaitGPUController(state.JID)
 
 			// Should kill process if still running since GPU controller might have exited prematurely
 			syscall.Kill(int(state.PID), syscall.SIGKILL)
