+++ pkg/api/server.go
@@ -1,20 +1,14 @@
 package api
 
 import (
-	"bytes"
 	"context"
 	"encoding/json"
 	"fmt"
-	"io"
 	"net"
-	"net/http"
 	"os"
-	"os/exec"
 	"strconv"
 	"strings"
 	"sync"
-	"syscall"
-	"time"
 
 	"github.com/mdlayher/vsock"
 	"github.com/swarnimarun/cadvisor/manager"
@@ -24,7 +18,6 @@ import (
 	"go.opentelemetry.io/otel/trace"
 
 	"github.com/cedana/cedana/pkg/api/runc"
-	"github.com/cedana/cedana/pkg/api/services/gpu"
 	task "github.com/cedana/cedana/pkg/api/services/task"
 	"github.com/cedana/cedana/pkg/db"
 	"github.com/cedana/cedana/pkg/jobservice"
@@ -35,22 +28,17 @@ import (
 	"github.com/spf13/viper"
 	"golang.org/x/sys/unix"
 	"google.golang.org/grpc"
-	"google.golang.org/grpc/credentials/insecure"
 	"google.golang.org/grpc/health"
 	healthcheckgrpc "google.golang.org/grpc/health/grpc_health_v1"
 	"google.golang.org/grpc/reflection"
-	"google.golang.org/protobuf/encoding/protojson"
-	"google.golang.org/protobuf/proto"
 )
 
 const (
-	DEFAULT_HOST                = "0.0.0.0"
-	PROTOCOL                    = "tcp"
-	CEDANA_CONTAINER_NAME       = "binary-container"
-	SERVER_LOG_MODE             = os.O_APPEND | os.O_CREATE | os.O_WRONLY
-	SERVER_LOG_PERMS            = 0o644
-	GPU_CONTROLLER_LOG_PATH     = "/tmp/cedana-gpucontroller.log"
-	GPU_CONTROLLER_WAIT_TIMEOUT = 10 * time.Second
+	DEFAULT_HOST          = "0.0.0.0"
+	PROTOCOL              = "tcp"
+	CEDANA_CONTAINER_NAME = "binary-container"
+	SERVER_LOG_MODE       = os.O_APPEND | os.O_CREATE | os.O_WRONLY
+	SERVER_LOG_PERMS      = 0o644
 )
 
 type service struct {
@@ -218,40 +206,12 @@ func StartServer(cmdCtx context.Context, opts *ServeOpts) error {
 			}
 		}
 
-		var wg sync.WaitGroup
 		if opts.GPUEnabled {
-			wg.Add(2)
-			go func() {
-				defer wg.Done()
-				gpuControllerPath := utils.GpuControllerBinaryPath
-				if s := viper.GetString("gpu_controller_path"); s != "" {
-					gpuControllerPath = s
-				}
-				log.Info().Str("gpu_controller_path", gpuControllerPath).Msg("Ensuring GPU Controller exists.")
-				err = pullGPUBinary(cmdCtx, utils.GpuControllerBinaryName, gpuControllerPath)
-				if err != nil {
-					log.Error().Err(err).Msg("could not pull gpu controller")
-					cancel(err)
-					return
-				}
-			}()
-
-			go func() {
-				defer wg.Done()
-				gpuSharedLibPath := utils.GpuSharedLibPath
-				if s := viper.GetString("gpu_shared_lib_path"); s != "" {
-					gpuSharedLibPath = s
-				}
-				log.Info().Str("gpu_shared_lib_path", gpuSharedLibPath).Msg("Ensuring LibCedana library exists.")
-				err = pullGPUBinary(cmdCtx, utils.GpuSharedLibName, gpuSharedLibPath)
-				if err != nil {
-					log.Error().Err(err).Msg("could not download libcedana")
-					cancel(err)
-					return
-				}
-			}()
-
-			wg.Wait()
+			err = DownloadGPUBinaries(cmdCtx)
+			if err != nil {
+				cancel(err)
+				return
+			}
 		}
 
 		log.Info().Str("host", DEFAULT_HOST).Uint32("port", opts.Port).Msg("server listening")
@@ -272,84 +232,6 @@ func StartServer(cmdCtx context.Context, opts *ServeOpts) error {
 	log.Debug().Msg("stopped RPC server gracefully")
 
 	return err
-}
-
-func (s *service) StartGPUController(ctx context.Context, uid, gid int32, groups []int32, out io.Writer, jid string) (*exec.Cmd, error) {
-	log.Debug().Int32("UID", uid).Int32("GID", gid).Ints32("Groups", groups).Msgf("starting gpu controller")
-	var gpuCmd *exec.Cmd
-	controllerPath := viper.GetString("gpu_controller_path")
-	if controllerPath == "" {
-		controllerPath = utils.GpuControllerBinaryPath
-	}
-	if _, err := os.Stat(controllerPath); os.IsNotExist(err) {
-		log.Error().Err(err).Send()
-		return nil, fmt.Errorf("no gpu controller at %s", controllerPath)
-	}
-
-	if viper.GetBool("gpu_debugging_enabled") {
-		controllerPath = strings.Join([]string{
-			"compute-sanitizer",
-			"--log-file /tmp/cedana-sanitizer.log",
-			"--print-level info",
-			"--leak-check=full",
-			controllerPath,
-		},
-			" ")
-	}
-
-	gpuCmd = exec.CommandContext(s.serverCtx, controllerPath, jid)
-	groupsUint32 := make([]uint32, len(groups))
-	for i, v := range groups {
-		groupsUint32[i] = uint32(v)
-	}
-	gpuCmd.SysProcAttr = &syscall.SysProcAttr{
-		Setsid: true,
-		Credential: &syscall.Credential{
-			Uid:    uint32(uid),
-			Gid:    uint32(gid),
-			Groups: groupsUint32,
-		},
-	}
-
-	if out != nil {
-		gpuCmd.Stderr = out
-		gpuCmd.Stdout = out
-	}
-
-	gpuCmd.Env = append(
-		os.Environ(),
-		"CEDANA_AUTH_TOKEN="+viper.GetString("connection.cedana_auth_token"),
-		"CEDANA_URL="+viper.GetString("connection.cedana_url"),
-	)
-
-	err := gpuCmd.Start()
-	if err != nil {
-		return nil, fmt.Errorf("could not start gpu controller %v", err)
-	}
-	log.Debug().Int("PID", gpuCmd.Process.Pid).Msgf("GPU controller starting...")
-
-	// poll gpu controller to ensure it is running
-	var opts []grpc.DialOption
-	var gpuConn *grpc.ClientConn
-	opts = append(opts, grpc.WithTransportCredentials(insecure.NewCredentials()))
-
-	gpuConn, err = grpc.NewClient("127.0.0.1:50051", opts...)
-	if err != nil {
-		return nil, fmt.Errorf("could not connect to gpu controller %v", err)
-	}
-	defer gpuConn.Close()
-
-	gpuServiceConn := gpu.NewCedanaGPUClient(gpuConn)
-
-	args := gpu.StartupPollRequest{}
-	waitCtx, _ := context.WithTimeout(ctx, GPU_CONTROLLER_WAIT_TIMEOUT)
-	resp, err := gpuServiceConn.StartupPoll(waitCtx, &args, grpc.WaitForReady(true))
-	if err != nil || !resp.Success {
-		return nil, fmt.Errorf("gpu controller did not start: %v", err)
-	}
-
-	log.Debug().Int("PID", gpuCmd.Process.Pid).Str("Log", GPU_CONTROLLER_LOG_PATH).Msgf("GPU controller started")
-	return gpuCmd, nil
 }
 
 func loggingStreamInterceptor() grpc.StreamServerInterceptor {
@@ -440,78 +322,6 @@ func (s *service) DetailedHealthCheck(ctx context.Context, req *task.DetailedHea
 	return resp, nil
 }
 
-func (s *service) GPUHealthCheck(
-	ctx context.Context,
-	req *task.DetailedHealthCheckRequest,
-	resp *task.DetailedHealthCheckResponse,
-) error {
-	gpuControllerPath := viper.GetString("gpu_controller_path")
-	if gpuControllerPath == "" {
-		gpuControllerPath = utils.GpuControllerBinaryPath
-	}
-
-	gpuSharedLibPath := viper.GetString("gpu_shared_lib_path")
-	if gpuSharedLibPath == "" {
-		gpuSharedLibPath = utils.GpuSharedLibPath
-	}
-
-	if _, err := os.Stat(gpuControllerPath); os.IsNotExist(err) {
-		resp.UnhealthyReasons = append(resp.UnhealthyReasons, fmt.Sprintf("gpu controller binary not found at %s", gpuControllerPath))
-	}
-
-	if _, err := os.Stat(gpuSharedLibPath); os.IsNotExist(err) {
-		resp.UnhealthyReasons = append(resp.UnhealthyReasons, fmt.Sprintf("gpu shared lib not found at %s", gpuSharedLibPath))
-	}
-
-	if len(resp.UnhealthyReasons) != 0 {
-		return nil
-	}
-
-	gpuOutBuf := &bytes.Buffer{}
-	gpuOut := io.MultiWriter(gpuOutBuf)
-	cmd, err := s.StartGPUController(ctx, req.UID, req.GID, req.Groups, gpuOut, "")
-	if err != nil {
-		log.Error().Err(err).Str("stdout/stderr", gpuOutBuf.String()).Msg("could not start gpu controller")
-		resp.UnhealthyReasons = append(resp.UnhealthyReasons, fmt.Sprintf("could not start gpu controller: %v", err))
-		return nil
-	}
-
-	defer func() {
-		err = cmd.Process.Kill()
-		if err != nil {
-			log.Error().Err(err).Msg("could not kill gpu controller")
-		} else {
-			log.Info().Int("PID", cmd.Process.Pid).Msgf("GPU controller killed")
-		}
-	}()
-
-	var opts []grpc.DialOption
-	opts = append(opts, grpc.WithTransportCredentials(insecure.NewCredentials()))
-
-	gpuConn, err := grpc.NewClient("127.0.0.1:50051", opts...)
-	if err != nil {
-		return err
-	}
-
-	defer gpuConn.Close()
-
-	gpuServiceConn := gpu.NewCedanaGPUClient(gpuConn)
-
-	args := gpu.HealthCheckRequest{}
-	gpuResp, err := gpuServiceConn.HealthCheck(ctx, &args)
-	if err != nil {
-		return err
-	}
-
-	if !gpuResp.Success {
-		resp.UnhealthyReasons = append(resp.UnhealthyReasons, "gpu health check did not return success")
-	}
-
-	resp.HealthCheckStats.GPUHealthCheck = gpuResp
-
-	return nil
-}
-
 func (s *service) GetConfig(ctx context.Context, req *task.GetConfigRequest) (*task.GetConfigResponse, error) {
 	resp := &task.GetConfigResponse{}
 	config, err := utils.GetConfig()
@@ -525,76 +335,4 @@ func (s *service) GetConfig(ctx context.Context, req *task.GetConfigRequest) (*t
 	}
 	resp.JSON = string(bytes)
 	return resp, nil
-}
-
-func pullGPUBinary(ctx context.Context, binary string, filePath string) error {
-	_, err := os.Stat(filePath)
-	if err == nil {
-		log.Info().Str("Path", filePath).Msgf("%s binary found, skipping download.", binary)
-		// TODO NR - check version and checksum of binary?
-		return nil
-	}
-	url := viper.GetString("connection.cedana_url") + "/k8s/gpu/" + binary
-	log.Info().Msgf("Downloading %s from %s", binary, url)
-
-	httpClient := &http.Client{}
-
-	req, err := http.NewRequestWithContext(ctx, "GET", url, nil)
-	if err != nil {
-		log.Err(err).Msg("failed to build http post request with jsonBody")
-		return err
-	}
-	req.Header.Set("Authorization", fmt.Sprintf("Bearer %s", viper.GetString("connection.cedana_auth_token")))
-	req.Header.Set("Content-Type", "application/json")
-
-	resp, err := httpClient.Do(req)
-	if err != nil || resp.StatusCode != http.StatusOK {
-		if resp != nil {
-			log.Error().Err(err).Int("status", resp.StatusCode).Msg("could not get gpu binary")
-		} else {
-			log.Error().Err(err).Msg("could not get gpu binary")
-		}
-		return fmt.Errorf("could not get gpu binary")
-	}
-	defer resp.Body.Close()
-
-	file, err := os.OpenFile(filePath, os.O_CREATE|os.O_WRONLY, 0755)
-	if err == nil {
-		err = os.Chmod(filePath, 0755)
-	}
-	if err != nil {
-		log.Err(err).Msg("could not create file")
-		return err
-	}
-	defer file.Close()
-
-	_, err = io.Copy(file, resp.Body)
-	if err != nil {
-		log.Err(err).Msg("could not read file from response")
-		return err
-	}
-	log.Info().Msgf("%s downloaded to %s", binary, filePath)
-	return err
-}
-
-func payloadToJSON(payload any) string {
-	if payload == nil {
-		return "null"
-	}
-
-	protoMsg, ok := payload.(proto.Message)
-	if !ok {
-		return fmt.Sprintf("%+v", payload)
-	}
-
-	marshaler := protojson.MarshalOptions{
-		EmitUnpopulated: true,
-		Indent:          "  ",
-	}
-	jsonData, err := marshaler.Marshal(protoMsg)
-	if err != nil {
-		return fmt.Sprintf("Error marshaling to JSON: %v", err)
-	}
-
-	return string(jsonData)
 }
